Move VerifySha256Checksum() to datasource package
