+++ pkg/artifactory/datasource/datasource.go
@@ -1,0 +1,26 @@
+package datasource
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+	"io"
+	"os"
+)
+
+func VerifySha256Checksum(path string, expectedSha256 string) (bool, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return false, err
+	}
+	defer func(f *os.File) {
+		_ = f.Close()
+	}(f)
+
+	hasher := sha256.New()
+
+	if _, err := io.Copy(hasher, f); err != nil {
+		return false, err
+	}
+
+	return hex.EncodeToString(hasher.Sum(nil)) == expectedSha256, nil
+}
