+++ pkg/utils/util.go
@@ -1,10 +1,7 @@
 package utils
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
-	"io"
 	"math/rand"
 	"net/http"
 	"net/url"
@@ -191,24 +188,6 @@ func FileExists(path string) bool {
 		}
 	}
 	return true
-}
-
-func VerifySha256Checksum(path string, expectedSha256 string) (bool, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return false, err
-	}
-	defer func(f *os.File) {
-		_ = f.Close()
-	}(f)
-
-	hasher := sha256.New()
-
-	if _, err := io.Copy(hasher, f); err != nil {
-		return false, err
-	}
-
-	return hex.EncodeToString(hasher.Sum(nil)) == expectedSha256, nil
 }
 
 type SupportedRepoClasses struct {
